+++ fcxref/cli.py
@@ -1,6 +1,7 @@
 import argparse
 import logging
 import os
+from pathlib import Path
 
 from fcxref.remove import remove
 
@@ -72,8 +73,7 @@ def main():
 
         def format_reference(reference: Reference) -> str:
             formatted_reference = str(reference)
-            word = 'direct' if str(property) == reference.match else 'indirect'
-            formatted_reference += ' {}'.format(word)
+            formatted_reference += ' {}'.format(get_reference_type(reference, property))
             return formatted_reference
 
         num_references = len(references)
@@ -160,6 +160,23 @@ def query_yes_no(question, default: str = 'yes'):
         else:
             print('Please respond with "yes" or "no" (or "y" or "n").\n')
 
+
+def get_reference_type(reference: Reference, property: Query) -> str:
+    if is_source(reference, property):
+        return 'source'
+    elif str(property) == reference.match:
+        return 'direct'
+    else:
+        return 'indirect'
+
+
+def is_source(reference: Reference, property: Query) -> bool:
+    return (
+        reference.reference_attribute == 'alias' and
+        reference.object_name == property.object_name and
+        property.matches_document(reference.document_path) and
+        reference.content == property.property_name
+    )
 
 if __name__ == '__main__':
     main()