+++ fcxref/find/make_find.py
@@ -3,7 +3,6 @@ import re
 from typing import Callable, Dict, List
 from xml.etree.ElementTree import Element
 
-from .extract_document import extract_document
 from .find_references_in_root import find_references_in_root
 from .query import Query
 from .reference import Reference
@@ -19,8 +18,7 @@ def make_find(find_root_by_document_path: Callable[[str], Dict[str, Element]]):
         logger.debug(f'Finding references in base path {base_path}')
         for document_path, root in root_by_document_path.items():
             logger.debug(f'Checking document {document_path}')
-            document = extract_document(document_path)
-            query_matches_document = document == query.document
+            query_matches_document = query.matches_document(document_path)
             pattern = r'\b{}\b'.format(query.property_name) if query_matches_document else query.to_regex()
             query_pattern = re.compile(pattern)
             references_in_document = find_references_in_root(